cleaned up main app dir by renaming rootReducer.js to reducers.js and merging routeReducer.js into reducers.js; removed export for routeReducer fn
