+++ app/reducers.js
@@ -1,0 +1,48 @@
+/**
+ * Combine all reducers in this file and export the combined reducers.
+ * If we were to do this in store.js, reducers wouldn't be hot reloadable.
+ */
+
+import { combineReducers } from 'redux-immutable';
+import globalReducer from 'App/reducer';
+import { fromJS } from 'immutable';
+import { LOCATION_CHANGE } from 'react-router-redux';
+
+/*
+ * routeReducer
+ *
+ * The reducer merges route location changes into our immutable state.
+ * The change is necessitated by moving to react-router-redux@4
+ *
+ */
+
+// Initial state
+const routeInitialState = fromJS({
+  locationBeforeTransitions: null,
+});
+
+/**
+ * merge route state into immutable
+ *
+ * @param {object} state = routeInitialState
+ * @param {object} action
+ * @returns {object}
+ */
+function routeReducer(state = routeInitialState, action) {
+  switch (action.type) {
+    case LOCATION_CHANGE:
+      return state.merge({
+        locationBeforeTransitions: action.payload,
+      });
+    default:
+      return state;
+  }
+}
+
+export default function createReducer(asyncReducers) {
+  return combineReducers({
+    route: routeReducer,
+    global: globalReducer,
+    ...asyncReducers,
+  });
+}
